+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   [#912](https://github.com/owncloud/music/pull/912) @PVince81
 - Option `rescan-modified` to the `occ` command `music:scan`
   [#843](https://github.com/owncloud/music/issues/843)
+- Stop button shown with long press on the play/pause button
+  [#911](https://github.com/owncloud/music/issues/911)
 
 ## 1.4.1 - 2021-10-31
 ### Added